[6366] Better seperate diagram layout tools
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6366
