+++ CHANGELOG.adoc
@@ -69,6 +69,7 @@ It now comes with small improvements in the filter bar for better reuse
 Switch all usage of the logger to the fluent API to prepare for the addition of more data.
 - https://github.com/eclipse-sirius/sirius-web/issues/6250[#6250] [selection] Move the selection dialog default settings to the backend
 Improve the log message used when logging exceptions.
+- https://github.com/eclipse-sirius/sirius-web/issues/6366[#6366] [diagram] Extract layout tools in several hooks.
